+++ packages/core/src/_test/e2e/factory/factory.test.ts
@@ -40,10 +40,7 @@ afterEach(() => {
 
 test("factory", async (context) => {
   const options = buildOptions({
-    cliOptions: {
-      rootDir: "./src/_test/e2e/factory",
-      configFile: "ponder.config.ts",
-    },
+    cliOptions: { root: "./src/_test/e2e/factory", config: "ponder.config.ts" },
   });
   const testOptions = {
     ...options,
@@ -60,12 +57,10 @@ test("factory", async (context) => {
   }
 
   const ponder = new Ponder({ options: testOptions });
-  await ponder.setup({
+  await ponder.start({
     syncStore: context.syncStore,
     indexingStore: context.indexingStore,
   });
-
-  await ponder.start();
 
   await onAllEventsIndexed(ponder);