+++ packages/core/src/sync-gateway/service.ts
@@ -1,10 +1,10 @@
-import Emittery from "emittery";
+import { Emittery } from "@/utils/emittery.js";
 import type { Hex } from "viem";
 
 import type { Common } from "@/Ponder.js";
 import type { Network } from "@/config/networks.js";
-import type { Source } from "@/config/sources.js";
 import { sourceIsFactory, sourceIsLogFilter } from "@/config/sources.js";
+import type { Source } from "@/config/sources.js";
 import type { SyncStore } from "@/sync-store/store.js";
 import {
   type Checkpoint,